+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Implement `Display` for `Value` and `Expr` types
 - Lalrpop parser for rules written in reval DSL
+- Implement `to_upper` and `to_lower` functions
 
 ### Changed
 - Return wrapped error type from json parsing rules